+++ PSFramework/changelog.md
@@ -1,7 +1,11 @@
 # CHANGELOG
 
 ## X.X.XX.XX : XXXX-XX-XX
+ - New: Add command Write-PSFMessageProxy (#81)
+ - New: Add command Set-PSFTypeAlias (#71)
  - Upd: Removing the module or closing the process will now stop all registered runspaces. This is designed to avoid hanging resources.
+ - Upd: Configuration: Added `Unchanged` property, in order to allow detection of settings that weren't changed by the user. (#79)
+ - Upd: Slight performance improvements on `Write-PSFMessage` and `Stop-PSFFunction`
  - Fix: Fixed critical concept error in Stop-PSFFunction, causing invalid termination in flowcontrol using commands. (#80)
 
 ## 0.9.11.25 : 2018-03-11